% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_app_file}
\alias{add_app_file}
\title{Add App}
\usage{
add_app_file()
}
\description{
Add the \code{app.R} required to run on RStudio Connect,
RStudio server, and shinyapps.io.
}
\examples{
\dontrun{add_app_file()}

}
