% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{add_module}
\alias{add_module}
\title{Add module}
\usage{
add_module(name)
}
\arguments{
\item{name}{Name of the module.
This is used to dynamically create the names of the
functions and file.}
}
\description{
Add boilerplate for a module.
}
\examples{
\dontrun{add_module("module_name")}

}
