% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{bootstrap_version}
\alias{bootstrap_version}
\title{Default Bootstrap Version}
\usage{
bootstrap_version()
}
\description{
Gets the default Bootstrap version based
on the Shiny version installed.
Shiny > 1.6 uses version 5 while
earlier version use Bootstrap 4.
}
