% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build}
\usage{
build()
}
\value{
Invisibly returns a boolean indicating
whether the operation was successful.
}
\description{
Runs the various scripts that were created with
\code{setup} functions.
}
\examples{
\dontrun{build()}

}
