% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_roclet}
\alias{build_roclet}
\title{Build Roclet}
\usage{
build_roclet()
}
\description{
Roclet to run \code{\link[=build]{build()}} when documenting.
Meant to be used in \code{DESCRIPTION}, e.g.:
\code{Roxygen: list(markdown = TRUE, roclets = c("namespace", "collate", "rd", "leprechaun::build_roclet"))}
}
