% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_assets}
\alias{create_assets}
\title{Create Assets}
\usage{
create_assets(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages to the console.}
}
\description{
Create assets function and file \code{assets.R}.
}
