% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_input_handlers}
\alias{create_input_handlers}
\title{Create Input Handlers}
\usage{
create_input_handlers(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages to the console.}
}
\description{
Create input handlers functions and file \code{inputs-handlers.R}.
}
\examples{
\dontrun{create_input_handlers()}

}
