% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_onload}
\alias{create_onload}
\title{Create onLoad}
\usage{
create_onload(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to pring messages to the console.}
}
\description{
Create \code{zzz.R} file containing \code{.onLoad} function.
}
