% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_utils}
\alias{create_utils}
\title{Create Utils}
\usage{
create_utils(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages to the console.}
}
\description{
Create \code{utils.R} file containing helper functions.
}
\examples{
\dontrun{create_utils()}

}
