% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_penguin_plot.R
\name{penguin_plot_server}
\alias{penguin_plot_server}
\title{penguin_plot Server}
\usage{
penguin_plot_server(id, penguins)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
penguin_plot Server
}
\keyword{internal}
