% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_penguin_plot.R
\name{penguin_plotUI}
\alias{penguin_plotUI}
\title{penguin_plot UI}
\usage{
penguin_plotUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
penguin_plot UI
}
\keyword{internal}
