% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_penguin_table.R
\name{penguin_table_server}
\alias{penguin_table_server}
\title{penguin_table Server}
\usage{
penguin_table_server(id, penguins)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
penguin_table Server
}
\keyword{internal}
