% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_penguin_table.R
\name{penguin_tableUI}
\alias{penguin_tableUI}
\title{penguin_table UI}
\usage{
penguin_tableUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
penguin_table UI
}
\keyword{internal}
