% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{plugin_packer}
\alias{plugin_packer}
\title{Packer Plugin}
\usage{
plugin_packer(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages.}
}
\description{
Setup a packer script to easily bundle the JavaScipt.
}
\note{
This requires a scaffold of packer already in place.
Also, this will require using \link{build}.
}
