% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{plugin_sass}
\alias{plugin_sass}
\title{Sass PLugin}
\usage{
plugin_sass(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages.}
}
\description{
Setup basic structure for sass and helper script for
bundling.
}
\note{
This will require using \link{build}.
}
