% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffold_leprechaun}
\alias{scaffold_leprechaun}
\title{Scaffold Leprechaun}
\usage{
scaffold_leprechaun(ui = c("navbarPage", "fluidPage"), overwrite = FALSE)
}
\arguments{
\item{ui}{Type of UI to use.}

\item{overwrite}{Whether to force overwrite all files.
This is not recommended, make sure you have save and/or
committed and checked that the files that will be overwritten
can be before proceeding with this option.}
}
\description{
Scaffolds a leprechaun project.
This must be run from within a package
and should only be run once per project.
}
