% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_packer}
\alias{setup_packer}
\title{Setup Packer}
\usage{
setup_packer()
}
\description{
Setup a packer script to easily bundle the JavaScipt.
}
\note{
This requires a scaffold of packer already in place.
}
