% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{sitrep}
\alias{sitrep}
\title{Sitrep}
\usage{
sitrep()
}
\value{
Invisibly returns a boolean indicating whether
the operation was successful.
}
\description{
Run a check on the project, prints useful
messages.
}
\examples{
\dontrun{sitrep()}

}
