% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_config}
\alias{use_config}
\title{Config}
\usage{
use_config(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages.}
}
\value{
Invisibly returns a boolean indicating whether
the operation was successful.
}
\description{
Setup a configuration file and helper functions.
}
