% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_endpoints_utils}
\alias{use_endpoints_utils}
\title{Endpoints}
\usage{
use_endpoints_utils(overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite existing files.}

\item{quiet}{Whether to print messages.}
}
\value{
Invisibly returns a boolean indicating whether
the operation was successful.
}
\description{
Add utility functions to create and serve
session-specific endpoints.
}
