% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_html_utils}
\alias{use_html_utils}
\title{Htmltools Utils}
\usage{
use_html_utils(overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite existing files.}

\item{quiet}{Whether to print messages.}
}
\value{
Invisibly returns a boolean indicating whether
the operation was successful.
}
\description{
Add htmltools utility functions, e.g.: shorthands to
create columns.
}
