% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_packer}
\alias{use_packer}
\title{Packer}
\usage{
use_packer(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages.}
}
\description{
Setup a packer script to easily bundle the JavaScipt.
}
\note{
This requires a scaffold of packer already in place.
Also, this will require using \link{build}.
}
