% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_sass}
\alias{use_sass}
\title{Sass}
\usage{
use_sass(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print messages.}
}
\description{
Setup basic structure for sass and helper script for
bundling.
}
\note{
This will require using \link{build}.
}
