% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_js_utils}
\alias{use_js_utils}
\title{JavaScript Utils}
\usage{
use_js_utils(overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite existing files.}

\item{quiet}{Whether to print messages.}
}
\description{
Add utility JavaScript function, e.g.: to hide or show
element in the DOM.
}
\note{
This requires \link{use_packer}.
Also, it will require using \link{build}.
}
\examples{
\dontrun{use_js_utils()}

}
